fix: use DOUBLE PRECISION in V58 migration (PostgreSQL-compatible)
H2 accepts DOUBLE but PostgreSQL requires DOUBLE PRECISION. Also
corrects the version comment from V57 to V58.

+++ backend/src/main/resources/db/migration/V58__add_dew_point_to_forecast_evaluation.sql
@@ -1,6 +1,6 @@
--- V57: Add dew_point_celsius to forecast_evaluation for mist/fog potential scoring.
+-- V58: Add dew_point_celsius to forecast_evaluation for mist/fog potential scoring.
 -- Captures dew point at 2 m above ground at the solar event slot.
 -- The gap between temperature_celsius and this value indicates fog/mist likelihood.
 -- NULL for historical rows fetched before this migration.
 
-ALTER TABLE forecast_evaluation ADD COLUMN IF NOT EXISTS dew_point_celsius DOUBLE;
+ALTER TABLE forecast_evaluation ADD COLUMN IF NOT EXISTS dew_point_celsius DOUBLE PRECISION;
